Got started on Unity unit tests and removed temporary project
