+++ RoboticonColony/Assets/Code/Classes/Inventory.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 53d04cbd1c2746641be919b7ac371901
-timeCreated: 1479996196
+guid: 439fe871292e296449e366f2873c0256
+timeCreated: 1480189643
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
