+++ RoboticonColony/Assets/Editor.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 5815d147368daab4292a360b20c0380a
+folderAsset: yes
+timeCreated: 1480190824
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
