+++ RoboticonColony/Assets/Editor/Unit_Tests/InventoryTest.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 4192b3bff47fc4a4aa995454565dcd78
+timeCreated: 1480191519
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
