+++ RoboticonColony/Assets/Editor/Unit_Tests/MarketTest.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 1c59f7bc40e80b84e9b3771c52c7f249
+timeCreated: 1480191156
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
