--- RoboticonColony/Assets/Inventory.cs
@@ -1,16 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-using System.Collections.Generic;
-
-public class Inventory
-{
-    public Inventory()
-    {
-
-    }
-
-    public void Exchange()
-    {
-
-    }
-}
