--- RoboticonColony/Assets/Resources.cs
@@ -1,9 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public struct Resources {
-    public int money;
-    public int ore;
-    public int power;
-    public int roboticon;
-}
