+++ RoboticonColony/Assets/Scenes.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: d5aad78c9c683a7499bc3d2599c2a094
+folderAsset: yes
+timeCreated: 1480189956
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
