+++ RoboticonColony/RoboticonColony.sln
@@ -3,6 +3,8 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2015
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RoboticonColony.CSharp", "RoboticonColony.CSharp.csproj", "{A4696F62-F605-20BE-129B-4E94DBDF0F7B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RoboticonColony.CSharp.Editor", "RoboticonColony.CSharp.Editor.csproj", "{C5E9515D-4E44-1382-D361-2A315B7271CB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -13,6 +15,10 @@ Global
 		{A4696F62-F605-20BE-129B-4E94DBDF0F7B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A4696F62-F605-20BE-129B-4E94DBDF0F7B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A4696F62-F605-20BE-129B-4E94DBDF0F7B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C5E9515D-4E44-1382-D361-2A315B7271CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C5E9515D-4E44-1382-D361-2A315B7271CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C5E9515D-4E44-1382-D361-2A315B7271CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C5E9515D-4E44-1382-D361-2A315B7271CB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
