--- TempRoboticonColony/TempRoboticonColony.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TempRoboticonColony", "TempRoboticonColony\TempRoboticonColony.csproj", "{3C9A08A7-97D5-4820-94E8-5D1969A8593E}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TempRoboticonColonyTest", "TempRoboticonColonyTest\TempRoboticonColonyTest.csproj", "{745E7CED-57F2-4D87-B5CA-1AFB30D4CD52}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3C9A08A7-97D5-4820-94E8-5D1969A8593E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3C9A08A7-97D5-4820-94E8-5D1969A8593E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3C9A08A7-97D5-4820-94E8-5D1969A8593E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3C9A08A7-97D5-4820-94E8-5D1969A8593E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{745E7CED-57F2-4D87-B5CA-1AFB30D4CD52}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{745E7CED-57F2-4D87-B5CA-1AFB30D4CD52}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{745E7CED-57F2-4D87-B5CA-1AFB30D4CD52}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{745E7CED-57F2-4D87-B5CA-1AFB30D4CD52}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
